src/requirements.txt
python-dotenv
azure-identity
agent-framework
agent-framework-devui